+++ CHANGELOG.md
@@ -9,10 +9,12 @@ All important changes to this project will be documented in this file.
 * Switched to using new Minecraft Wiki API
 * Added translation strings for death messages up to version `1.21.3`
 * Enhanced algorithm to get average loading server time
+* Added options to enable proxy for RSS and other requests separately
+* Added option to enable using User Agent from Chrome or Firefox when checking RSS feed instead of default one - `requests/1.31` or `feedparser/6.0.10`
 
 #### Fixed:
 
-* Added using proxy for getting RSS feed
+* Fixed proxy for HTTPS connections
 
 #### Breaking changes: